--- src/components/RoleFormes/investor/investorForm.module.scss
@@ -1,3 +1,0 @@
-@import "../../../sass/utils/container";
-@import "../../../sass/utils/mixins";
-@import "../../../sass/utils/variables";
