+++ src/components/UI/Forms/RoleFormes/partner/partnerForm.module.scss
@@ -1,0 +1,3 @@
+@import '../../../../../sass/utils/container';
+@import '../../../../../sass/utils/mixins';
+@import '../../../../../sass/utils/variables';
