+++ src/libs/enums/helpers/getRoleForm.tsx
@@ -1,9 +1,9 @@
-import JuniorForm from "@/components/RoleFormes/junior/JuniorForm";
-import PartnerForm from "@/components/RoleFormes/partner/PartnerForm";
-import { RoleType } from "@/utils/types/Role";
-import { Roles } from "../app/Role";
-import MentorForm from "@/components/RoleFormes/mentor/MentorForm";
-import InvestorForm from "@/components/RoleFormes/investor/InvestorForm";
+import JuniorForm from '@/components/UI/Forms/RoleFormes/junior/JuniorForm';
+import PartnerForm from '@/components/UI/Forms/RoleFormes/partner/PartnerForm';
+import { RoleType } from '@/utils/types/Role';
+import { Roles } from '../app/Role';
+import MentorForm from '@/components/UI/Forms/RoleFormes/mentor/MentorForm';
+import InvestorForm from '@/components/UI/Forms/RoleFormes/investor/InvestorForm';
 
 export const getRoleForm = (role: RoleType): React.ReactElement => {
   console.log(Roles.JUNIOR);
